requirements.txt
seleniumbase
playwright
gspread
python-dotenv
pytest
requests
google-auth
google-auth-oauthlib
google-auth-httplib2